+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## master (unreleased)
 
+* Allow the disabling of all cops from `config/rubocop.yml` with the `--disable-default-cops` option. ([@rickmzp][])
+
 ### Bugs fixed
 
 * [#1197](https://github.com/bbatsov/rubocop/issues/1197): Fix false positive for new lambda syntax in `SpaceInsideBlockBraces`. ([@jonas054][])
@@ -1023,3 +1025,4 @@
 [@ggilder]: https://github.com/ggilder
 [@salbertson]: https://github.com/salbertson
 [@camilleldn]: https://github.com/camilleldn
+[@rickmzp]: https://github.com/rickmzp